PILi_Quarts_V4.0_Studio/backend/requirements.txt
# PILi Quarts Backend - Requirements
# Enterprise-grade Python dependencies
# Following python-patterns and deployment-procedures skills

# ============================================================================
# CORE FRAMEWORK
# ============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ============================================================================
# DATABASE
# ============================================================================
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# ============================================================================
# VALIDATION & SERIALIZATION
# ============================================================================
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ============================================================================
# AI & ML
# ============================================================================
google-generativeai==0.3.2
tenacity==8.2.3

# ============================================================================
# DOCUMENT GENERATION
# ============================================================================
# PDF Generation
reportlab==4.0.9

# Word Generation
python-docx==1.1.0
htmldocx==0.0.6

# Excel Generation
openpyxl==3.1.2

# ============================================================================
# HTTP & NETWORKING
# ============================================================================
httpx==0.26.0

websockets==12.0
python-socketio==5.11.0

# ============================================================================
# CACHING & SESSION (Optional but recommended)
# ============================================================================
redis==5.0.1
aioredis==2.0.1

# ============================================================================
# UTILITIES
# ============================================================================
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.3

# ============================================================================
# LOGGING & MONITORING
# ============================================================================
structlog==24.1.0

# ============================================================================
# TESTING
# ============================================================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# ============================================================================
# CODE QUALITY
# ============================================================================
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# ============================================================================
# DEVELOPMENT
# ============================================================================
ipython==8.20.0
